# statistics_manager_service/models/instant_voltage_metric.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from statistics_manager_service.models.base_model import Model
from statistics_manager_service.models.voltage_units import VoltageUnits
from statistics_manager_service import util

from statistics_manager_service.models.voltage_units import VoltageUnits  # noqa: E501

class InstantVoltageMetric(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, user_id=None, timestamp=None, value=None, units=None):  # noqa: E501
        """InstantVoltageMetric - a model defined in OpenAPI

        :param user_id: The user_id of this InstantVoltageMetric.  # noqa: E501
        :type user_id: str
        :param timestamp: The timestamp of this InstantVoltageMetric.  # noqa: E501
        :type timestamp: datetime
        :param value: The value of this InstantVoltageMetric.  # noqa: E501
        :type value: float
        :param units: The units of this InstantVoltageMetric.  # noqa: E501
        :type units: VoltageUnits
        """
        self.openapi_types = {
            'user_id': str,
            'timestamp': datetime,
            'value': float,
            'units': VoltageUnits
        }

        self.attribute_map = {
            'user_id': 'user_id',
            'timestamp': 'timestamp',
            'value': 'value',
            'units': 'units'
        }

        self._user_id = user_id
        self._timestamp = timestamp
        self._value = value
        self._units = units

    @classmethod
    def from_dict(cls, dikt) -> 'InstantVoltageMetric':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The InstantVoltageMetric of this InstantVoltageMetric.  # noqa: E501
        :rtype: InstantVoltageMetric
        """
        return util.deserialize_model(dikt, cls)

    @property
    def user_id(self) -> str:
        """Gets the user_id of this InstantVoltageMetric.


        :return: The user_id of this InstantVoltageMetric.
        :rtype: str
        """
        return self._user_id

    @user_id.setter
    def user_id(self, user_id: str):
        """Sets the user_id of this InstantVoltageMetric.


        :param user_id: The user_id of this InstantVoltageMetric.
        :type user_id: str
        """
        if user_id is None:
            raise ValueError("Invalid value for `user_id`, must not be `None`")  # noqa: E501
        if user_id is not None and len(user_id) > 10:
            raise ValueError("Invalid value for `user_id`, length must be less than or equal to `10`")  # noqa: E501
        if user_id is not None and len(user_id) < 10:
            raise ValueError("Invalid value for `user_id`, length must be greater than or equal to `10`")  # noqa: E501

        self._user_id = user_id

    @property
    def timestamp(self) -> datetime:
        """Gets the timestamp of this InstantVoltageMetric.


        :return: The timestamp of this InstantVoltageMetric.
        :rtype: datetime
        """
        return self._timestamp

    @timestamp.setter
    def timestamp(self, timestamp: datetime):
        """Sets the timestamp of this InstantVoltageMetric.


        :param timestamp: The timestamp of this InstantVoltageMetric.
        :type timestamp: datetime
        """
        if timestamp is None:
            raise ValueError("Invalid value for `timestamp`, must not be `None`")  # noqa: E501

        self._timestamp = timestamp

    @property
    def value(self) -> float:
        """Gets the value of this InstantVoltageMetric.


        :return: The value of this InstantVoltageMetric.
        :rtype: float
        """
        return self._value

    @value.setter
    def value(self, value: float):
        """Sets the value of this InstantVoltageMetric.


        :param value: The value of this InstantVoltageMetric.
        :type value: float
        """
        if value is None:
            raise ValueError("Invalid value for `value`, must not be `None`")  # noqa: E501
        if value is not None and value < 0:  # noqa: E501
            raise ValueError("Invalid value for `value`, must be a value greater than or equal to `0`")  # noqa: E501

        self._value = value

    @property
    def units(self) -> VoltageUnits:
        """Gets the units of this InstantVoltageMetric.


        :return: The units of this InstantVoltageMetric.
        :rtype: VoltageUnits
        """
        return self._units

    @units.setter
    def units(self, units: VoltageUnits):
        """Sets the units of this InstantVoltageMetric.


        :param units: The units of this InstantVoltageMetric.
        :type units: VoltageUnits
        """
        if units is None:
            raise ValueError("Invalid value for `units`, must not be `None`")  # noqa: E501

        self._units = units