# statistics_manager_service/models/voltage_metrics.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from statistics_manager_service.models.base_model import Model
from statistics_manager_service.models.voltage_metric import VoltageMetric
from statistics_manager_service import util

from statistics_manager_service.models.voltage_metric import VoltageMetric  # noqa: E501

class VoltageMetrics(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, user_id=None, metrics=None):  # noqa: E501
        """VoltageMetrics - a model defined in OpenAPI

        :param user_id: The user_id of this VoltageMetrics.  # noqa: E501
        :type user_id: str
        :param metrics: The metrics of this VoltageMetrics.  # noqa: E501
        :type metrics: List[VoltageMetric]
        """
        self.openapi_types = {
            'user_id': str,
            'metrics': List[VoltageMetric]
        }

        self.attribute_map = {
            'user_id': 'user_id',
            'metrics': 'metrics'
        }

        self._user_id = user_id
        self._metrics = metrics

    @classmethod
    def from_dict(cls, dikt) -> 'VoltageMetrics':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The VoltageMetrics of this VoltageMetrics.  # noqa: E501
        :rtype: VoltageMetrics
        """
        return util.deserialize_model(dikt, cls)

    @property
    def user_id(self) -> str:
        """Gets the user_id of this VoltageMetrics.


        :return: The user_id of this VoltageMetrics.
        :rtype: str
        """
        return self._user_id

    @user_id.setter
    def user_id(self, user_id: str):
        """Sets the user_id of this VoltageMetrics.


        :param user_id: The user_id of this VoltageMetrics.
        :type user_id: str
        """
        if user_id is None:
            raise ValueError("Invalid value for `user_id`, must not be `None`")  # noqa: E501
        if user_id is not None and len(user_id) > 10:
            raise ValueError("Invalid value for `user_id`, length must be less than or equal to `10`")  # noqa: E501
        if user_id is not None and len(user_id) < 10:
            raise ValueError("Invalid value for `user_id`, length must be greater than or equal to `10`")  # noqa: E501

        self._user_id = user_id

    @property
    def metrics(self) -> List[VoltageMetric]:
        """Gets the metrics of this VoltageMetrics.


        :return: The metrics of this VoltageMetrics.
        :rtype: List[VoltageMetric]
        """
        return self._metrics

    @metrics.setter
    def metrics(self, metrics: List[VoltageMetric]):
        """Sets the metrics of this VoltageMetrics.


        :param metrics: The metrics of this VoltageMetrics.
        :type metrics: List[VoltageMetric]
        """
        if metrics is None:
            raise ValueError("Invalid value for `metrics`, must not be `None`")  # noqa: E501

        self._metrics = metrics